requirements.txt
streamlit
sentence-transformers
faiss-cpu
tiktoken
numpy
python-dotenv
requests